Initial fetch module complete with simple testbench
I completed the skeleton for the fetch module. I believe it satisfies
the requirements for now. It just needs to be tied in with the decoder
module. There is a need to create an overall test bench so we can
incorporate the parser. The parse will probably need to be reworked
slightly to make this work.