PD2 pretty well complete - just needs more testing
Made some major changes to how the structure of the code was laid out.
Now there is an overall module that contains all the components needed
for the CPU - processor.v. The processor_tb.v is the test bench that
tests the entire system. We will use this to test the fetch and decode
module because this test bench incorporates the SREC parser. Everything
seems to be working, we just need to do some testing to make sure the
decoder is doing exactly what it is suppose to and match it with the
instructions in benchmark program