Decode stage has register file
Got the register file instantiated into the decode stage. It seems that
everything is working and the values are correct in the registers. I
fixed the bug where we execute the first instruction twice. I am
assuming that the decode stage has the next PC and NOT the PC of the
instruction in the IR (instruction register)