Tested all bench marks except CheckVowel
I tested each of the bench marks for the expected output and all the
benchmarks passed. I did not test CheckVowel yet because I am still
trying to figure out the register that holds the value of j. I had a bug
in my branch resolve module that I fixed. The SRA instruction control
was also not implemented so I did that. There however is a problem with
SRA through I believe because it is not filling the upper bits with 1's
when we have a negative number. It did not affect the benchmarks but we
should still fix it. Also added the times to start and finish execution
for each of the benchmarks in the TCL script.